requirements.txt
transformers
huggingface_hub
hydra-core
pyJoules
datasets
tqdm
torch
pynvml
